+++ lib/pages/more/MoreMenu.dart
@@ -6,7 +6,9 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:app/assets/icons/customIcons.dart';
 
 import 'package:app/assets/components/customTile.dart';
-import 'package:app/assets/routes/customLaunch.dart';
+
+import 'package:app/locator.dart';
+import 'package:app/services/navigation_service.dart';
 
 import 'package:app/pages/more/ProfileTile.dart';
 import 'package:app/routes.dart';
@@ -24,6 +26,10 @@ final profileTiles = <Map>[
   //  'page': Routes.profile,
   //},
   {'sectionHeading': "The app"},
+  {
+    'profileTile': ProfileTile("About Us",CustomIcons.ic_faq),
+    'link': "https://now-u.com/aboutus",
+  },
   {
     'profileTile': ProfileTile("Our partners", CustomIcons.ic_partners),
     'page': Routes.parteners,
@@ -86,8 +92,15 @@ final profileTiles = <Map>[
     'external': true
   },
 ];
-
+///The More page ![More Page](https://i.ibb.co/xDHyMPj/slack.png)
+///
+/// This Widget takes in the [profileTiles] and goes over it
+/// checking if the elements inside it is either sectionHeading
+/// or [ProfileTile]
 class Profile extends StatelessWidget {
+  
+  final NavigationService _navigationService = locator<NavigationService>();
+
   @override
   Widget build(BuildContext context) {
     return ViewModelBuilder<BaseModel>.reactive(
@@ -134,8 +147,8 @@ class Profile extends StatelessWidget {
                                       context, profileTiles[index]["page"]);
                                 } else if (profileTiles[index]["link"] !=
                                     null) {
-                                  customLaunch(
-                                      context, profileTiles[index]["link"],
+                                  _navigationService.launchLink(
+                                      profileTiles[index]["link"],
                                       isExternal: profileTiles[index]
                                               ['external'] ??
                                           false);
@@ -170,7 +183,6 @@ class Profile extends StatelessWidget {
                     SizedBox(height: 20),
 
                     // Dev Tools
-                    // TODO add this back in its super duper handy
                     model.currentUser.isStagingUser()
                       ? Column(
                           children: [
@@ -186,6 +198,12 @@ class Profile extends StatelessWidget {
                               child: ProfileTile("Log out", FontAwesomeIcons.code),
                               onTap: () {
                                 model.logout();
+                              }
+                            ),
+                            GestureDetector(
+                              child: ProfileTile("Navigate test", FontAwesomeIcons.code),
+                              onTap: () {
+                                _navigationService.launchLink("internal:learningTopic&id=-1");
                               }
                             ),
                           ]