+++ lib/services/dialog_service.dart
@@ -1,16 +1,33 @@
 import 'dart:async';
+import 'package:app/assets/components/darkButton.dart';
 
 class AlertRequest {
   final String title;
   final String description;
+  final List buttons;
 
-  AlertRequest({this.title, this.description});
+  AlertRequest({this.title, this.description, this.buttons});
 }
 
-
 class AlertResponse {
-  final bool confirmed;
-  AlertResponse({this.confirmed});
+  final dynamic response;
+  AlertResponse({this.response});
+}
+
+class DialogButton {
+  String text;
+  dynamic response; // If the button is clicked what should the dialog service return 
+  DarkButtonStyle style;
+  //Function onClick; // This might be handy in the future
+  bool closeOnClick;
+  
+  DialogButton({
+    this.text,
+    this.response,
+    this.style,
+    this.closeOnClick,
+    //this.onClick,
+  });
 }
 
 
@@ -25,9 +42,10 @@ class DialogService {
   Future showDialog({
     String title,
     String description,
+    List buttons,
   }) {
     _dialogCompleter = Completer<AlertResponse>();
-    _showDialogListener(AlertRequest(title: title, description: description));
+    _showDialogListener(AlertRequest(title: title, description: description, buttons: buttons));
     return _dialogCompleter.future;
   }